Fix compiler warnings (v1.8.2)
- Fixed missing-field-initializers for all struct initializers
- Fixed reorder warnings in CryptoEmulator and VirtualSecurityChip
- Fixed type-limits warning in CryptoUtils
- Fixed unused-variable warning in ProfileManager
- Fixed sign-compare warning in RealisticProfileGenerator
- Added warning suppressions to CMakeLists.txt
- All 30 source files compile with 0 errors, 0 warnings

+++ src/RealisticProfileGenerator.cpp
@@ -1226,7 +1226,7 @@ std::vector<DeviceProfile> RealisticProfileGenerator::generateUniqueBatch(int co
     std::vector<DeviceProfile> profiles;
     std::set<std::string> usedSerials;
     
-    for (int i = 0; i < count * 2 && profiles.size() < count; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(count) * 2 && profiles.size() < static_cast<size_t>(count); ++i) {
         auto result = generateRandomProfile(manufacturer);
         if (result.success && usedSerials.find(result.profile.serialNumber) == usedSerials.end()) {
             usedSerials.insert(result.profile.serialNumber);
